Adding refactor
